requirements.txt
asgiref==3.8.1
Django==5.1.7
djangorestframework==3.15.2
Faker==37.1.0
sqlparse==0.5.3
tzdata==2025.2
django-extensions
